Add files via upload
The code is done by python 3 and keras library. The code is run by google colab . First we had to upload FER 2013 dataset on to our drive . Then we generated a pre trained data model Emotion_little_vgg.h5 which is consist if trained data of five classes of FER-2013.